Added reusability & consistency for Workflow headers on: Workflow Invocations Page and Workflow Run Form Page
